requirements.txt
# ─── Core ─────────────────────────────────────────────────────────────────────
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0

# ─── Math / Symbolic ──────────────────────────────────────────────────────────
sympy>=1.12

# ─── Visualisation ────────────────────────────────────────────────────────────
matplotlib>=3.7.0

# ─── Knowledge Graph / Web scraping ───────────────────────────────────────────
requests>=2.31.0
beautifulsoup4>=4.12.0
mwparserfromhell>=0.6.5
SPARQLWrapper>=2.0.0

# ─── ML extras ────────────────────────────────────────────────────────────────
scikit-learn>=1.3.0

# ─── Dev / test ───────────────────────────────────────────────────────────────
pytest>=7.4.0
pytest-cov>=4.1.0

# ─── Optional: text embeddings ────────────────────────────────────────────────
# sentence-transformers>=2.2.0
